pyproject.toml
[tool.poetry]
name = "notes"
version = "0.1.0"
description = ""
authors = ["Nikita Smykov <132088678+Apicqq@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.2"
alembic = "^1.13.2"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
black = "^24.8.0"
pydantic-settings = "^2.4.0"
uvicorn = {extras = ["standart"], version = "^0.30.6"}
aiosqlite = "^0.20.0"
requests = "^2.32.3"
asyncpg = "^0.29.0"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
mixer = "^7.2.2"
httpx = "^0.27.2"
fastapi-pagination = "^0.12.26"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
